[bento-core] Detach pricing fetchers
