+++ packages/bento-core/src/pricings/Currency.ts
@@ -1,0 +1,1 @@
+export type Currency = 'usd';
